admin panel now only visible to users with role admin, admin routes also protected

+++ server/routes.py
@@ -114,7 +114,7 @@ def save_system_settings(name):
                     f":{request.remote_addr}")
     try:
         user = DBUtils.get_user(get_jwt_identity())
-        if DBUtils.user_isadmin(user.id):
+        if DBUtils.user_isadmin(user["id"]):
             try:
                 document = DBUtils.get_document(user_id="sidekick", name=name,
                                                 type=DOCTYPE_SYSTEM_SETTINGS)
@@ -130,11 +130,12 @@ def save_system_settings(name):
                 status=200,
                 mimetype='application/json'
             )
-        return app.response_class(
-            response=json.dumps({"success": False}),
-            status=403,
-            mimetype='application/json'
-        )
+        else:
+            return app.response_class(
+                response=json.dumps({"success": False}),
+                status=403,
+                mimetype='application/json'
+            )
     except Exception as e:
         log_exception(e)
         return str(e), 500
@@ -651,26 +652,32 @@ def reset_password():
     app.logger.info(
         f"/reset_password by user_id: {acting_user_id} for user_id: {user_id} "
         f"[POST] request from: {request.remote_addr}")
-    try:
-        result = DBUtils.reset_password(acting_user_id=acting_user_id,
-                                        user_id=user_id,
-                                        new_password=new_password)
-        if result['success']:
-            access_token = create_access_token(identity=data['user_id'])
-            result['access_token'] = access_token
-            return jsonify(result)
-        else:
-            return app.response_class(
-                response=json.dumps({"success": False, "message": result['message']}),
-                status=403,
-                mimetype='application/json'
-            )
-    except Exception as e:
-        app.logger.error(f"/reset_password user_id:{data['user_id']} "
-                         f"error:{str(e)}")
-        log_exception(e)
-        return jsonify({'success': False, 'message': str(e)})
-
+    if DBUtils.user_isadmin(acting_user_id):
+        try:
+            result = DBUtils.reset_password(acting_user_id=acting_user_id,
+                                            user_id=user_id,
+                                            new_password=new_password)
+            if result['success']:
+                access_token = create_access_token(identity=data['user_id'])
+                result['access_token'] = access_token
+                return jsonify(result)
+            else:
+                return app.response_class(
+                    response=json.dumps({"success": False, "message": result['message']}),
+                    status=403,
+                    mimetype='application/json'
+                )
+        except Exception as e:
+            app.logger.error(f"/reset_password user_id:{data['user_id']} "
+                            f"error:{str(e)}")
+            log_exception(e)
+            return jsonify({'success': False, 'message': str(e)})
+    else:
+        return app.response_class(
+            response=json.dumps({"success": False, "message": "Only admins can reset passwords"}),
+            status=403,
+            mimetype='application/json'
+        )
 
 @app.route('/delete_user', methods=['POST'])
 @jwt_required()
@@ -682,8 +689,13 @@ def delete_user():
     app.logger.info(
         f"/delete_user by acting_user_id: {acting_user_id} of user_id: {user_id_to_delete} [POST] request from"
         f":{request.remote_addr}")
+    if acting_user_id != user_id_to_delete and not DBUtils.user_isadmin(acting_user_id):
+        return app.response_class(
+            response=json.dumps({"success": False, "message": "Only admins can delete other users"}),
+            status=403,
+            mimetype='application/json'
+        )
     try:
-        # TODO: Check acting_user_id has admin role
         if DBUtils.login(acting_user_id, password)['success']:
             result = DBUtils.delete_user(user_id_to_delete)
             return jsonify(result)
